sentiment_analysis_agent/requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.2
httpx>=0.24.1
pydantic>=2.3.0
python-dotenv>=1.0.0
crewai>=0.27.0 